backend/requirements.txt
fastapi
uvicorn
pydantic
pytesseract
pdf2image
Pillow
chromadb
openai
tqdm
python-multipart
sentence-transformers 